Refactor PortForward server methods into the portforward package
